Move `TokenMap` to its own file
